Rename InstanceChain to Chain
