+++ CMakeLists.txt
@@ -30,9 +30,9 @@ set( ADMUTIL_ROOT ${CMAKE_CURRENT_SOURCE_DIR} )
 set( ADMUTIL_SOURCES
 		src/Platform.hpp
 		src/Precompiled.hpp
+		src/Containers/Chain.hpp
 		src/Containers/Dictionary.hpp
 		src/Containers/Dictionary.cpp
-		src/Containers/InstanceChain.hpp
 		src/Containers/Singleton.hpp 
 		src/Maths/Lerp.hpp 
 		src/Maths/Vec3.hpp
